+++ src/api/sdk/advanced-token/advanced-token-models.ts
@@ -1,0 +1,10 @@
+export namespace AdvancedTokenExchangeModels {
+    export interface AdvancedTokenExchangeRequest {
+        subject_token: string;
+        subject_token_type: string;
+        grant_type: string;
+        resource: string;
+    }
+
+    export interface AdvancedTokenExchangeResponse {}
+}